LayoutTests/fast/events/touch/ios/do-not-show-keyboard-after-preventing-touchstart-expected.txt


Verifies that the keyboard does not show up even after preventing default on touchstart when a regular input field is focused. To manually test, tap the red box; the keyboard should not appear.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS hasInputSession is false
PASS document.activeElement is textarea
PASS successfullyParsed is true

TEST COMPLETE
